Rename manager package into evaluator
